+++ integration/backward_compatibility_test.go
@@ -1,6 +1,9 @@
 package main
 
 import (
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"testing"
 	"time"
 
@@ -21,7 +24,7 @@ const (
 
 func TestBackwardCompatibilityWithChunksStorage(t *testing.T) {
 	s, err := e2e.NewScenario()
-	defer s.Close() // lint:ignore SA5001
+	defer s.Close() //nolint
 	require.NoError(t, err)
 
 	// Start dependencies.
@@ -30,7 +33,11 @@ func TestBackwardCompatibilityWithChunksStorage(t *testing.T) {
 	require.NoError(t, s.WaitReady("consul", "dynamodb"))
 
 	// Start Cortex components (ingester running on previous version).
-	require.NoError(t, writeCortexConfigFile(s))
+	require.NoError(t, ioutil.WriteFile(
+		filepath.Join(s.SharedDir(), cortexSchemaConfigFile),
+		[]byte(cortexSchemaConfigYaml),
+		os.ModePerm),
+	)
 	require.NoError(t, s.StartService(e2ecortex.NewTableManager("table-manager", ChunksStorage, "")))
 	require.NoError(t, s.StartService(e2ecortex.NewIngester("ingester-1", ChunksStorage, previousVersionImage)))
 	require.NoError(t, s.StartService(e2ecortex.NewDistributor("distributor", ChunksStorage, "")))
